requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
drf-api-logger>=1.1.14
python-dotenv>=1.0.0 